moved ActivitiesTreeView to a new file
